7 - Rename of Folder and Fix of Imports
