+++ number_trivia_app/lib/core/network/network_info.dart
