+++ number_trivia_app/lib/features/number_trivia/data/repositories/number_trivia_repository_impl.dart
@@ -4,7 +4,7 @@ import 'package:number_trivia_app/core/error/exceptions.dart';
 
 import '../../../../core/error/failures.dart';
 import '../../../../core/error/exceptions.dart';
-import '../../../../core/platform/network_info.dart';
+import '../../../../core/network/network_info.dart';
 import '../../domain/entities/number_trivia.dart';
 import '../../domain/repositories/number_trivia_repository.dart';
 import '../datasources/number_trivia_local_data_source.dart';
