+++ number_trivia_app/test/features/number_trivia/data/repositories/number_trivia_repository_impl_test.dart
@@ -3,7 +3,7 @@ import 'package:mockito/mockito.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:number_trivia_app/core/error/exceptions.dart';
 import 'package:number_trivia_app/core/error/failures.dart';
-import 'package:number_trivia_app/core/platform/network_info.dart';
+import 'package:number_trivia_app/core/network/network_info.dart';
 import 'package:number_trivia_app/features/number_trivia/data/datasources/number_trivia_local_data_source.dart';
 import 'package:number_trivia_app/features/number_trivia/data/datasources/number_trivia_remote_datasource.dart';
 import 'package:number_trivia_app/features/number_trivia/data/models/number_trivia_model.dart';
